requirements.txt
Django==5.0.0
pillow==10.1.0
qrcode==7.4.2
openpyxl==3.1.2

